requirements.txt
# Core dependencies
prefect>=2.0.0
pandas>=1.3.5
numpy>=1.21.5
sqlalchemy>=1.4.39

# AWS dependencies
prefect-aws>=0.3.0
boto3>=1.24.71
s3fs>=2022.8.2

# Snowflake dependencies
prefect-snowflake>=0.3.0
snowflake-sqlalchemy>=1.4.4
snowflake-connector-python>=2.7.12

# Environment and configuration
python-dotenv>=0.20.0
pyyaml>=6.0

# Development dependencies
black>=22.8.0
isort>=5.10.1
pytest>=7.1.3
pytest-cov>=4.0.0 